+++ docs/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.2] - ???
+### Changed
+- Fix an issue with FlawSource validation for sources that can be both
+  public and private (OSIDB-450)
+
 ## [2.2.1] - 2022-09-07
 ### Changed
 - Fix an issue with CVSSv3 validation that was preventing some flaws from